requirements.txt
google-generativeai>=0.8.0
streamlit>=1.38.0
matplotlib>=3.8.0
numpy>=1.26.0
python-dotenv>=1.0.0
livekit-api>=0.7.0
aiohttp>=3.9.0
